requirements.txt
aiohttp==3.5.4
async-generator==1.10
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.8.0
boto3==1.9.199
botocore==1.12.199
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
clarifai==2.6.2
clashroyale==4.0.1
configparser==3.7.4
dblpy==0.3.1
discord.py==1.2.3
docutils==0.14
fortnite==0.0.8
future==0.17.1
fuzzywuzzy==0.17.0
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.6.0
googletrans==2.4.0
grpcio==1.22.0
httplib2==0.13.1
idna==2.8
inflect==2.1.0
jaraco.itertools==4.4.2
jmespath==0.9.4
jsonschema==2.6.0
more-itertools==7.2.0
multidict==4.5.2
praw==6.3.1
prawcore==1.0.1
protobuf==3.9.0
psutil==5.6.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
Pybooru==4.1.0
pycparser==2.19
PyNaCl==1.3.0
python-box==3.4.2
python-dateutil==2.8.0
ratelimiter==1.2.0.post0
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
six==1.12.0
soupsieve==1.9.2
update-checker==0.16
uritemplate==3.0.0
urllib3==1.25.3
websocket-client==0.56.0
websockets==6.0
wolframalpha==3.0.1
xmltodict==0.12.0
yarl==1.3.0
youtube-dl==2019.8.2
